requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.0.0=pypi_0
albumentations=1.3.1=pypi_0
anyio=4.0.0=pypi_0
appnope=0.1.3=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.4.0=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.1.0=pypi_0
awscli=1.29.64=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
black=23.10.1=pypi_0
blas=1.0=openblas
bleach=6.0.0=pypi_0
botocore=1.31.64=pypi_0
bottleneck=1.3.5=py311ha0d4635_0
brotli=1.1.0=hb547adb_0
brotli-bin=1.1.0=hb547adb_0
brotlipy=0.7.0=py311h80987f9_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.1=hb547adb_0
ca-certificates=2023.7.22=hf0a4a13_0
cachetools=5.3.2=pypi_0
cairo=1.16.0=h29d4eff_2
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py311h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.4=pypi_0
comm=0.1.4=pypi_0
contourpy=1.0.5=py311h48ca7d4_0
cryptography=41.0.3=py311hd4332d6_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
docutils=0.16=pypi_0
eigen=3.4.0=hc021e02_0
executing=1.2.0=pypi_0
fastjsonschema=2.18.0=pypi_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.9.0=py311hca03da5_0
fontconfig=2.13.94=heb65262_0
fonttools=4.25.0=pyhd3eb1b0_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h1192e45_0
gettext=0.21.0=h13f89a0_1
giflib=5.2.1=h80987f9_3
glib=2.69.1=h514c7bf_2
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py311h40f64dc_0
gnutls=3.6.15=h887c41c_0
google-auth=2.23.3=pypi_0
google-auth-oauthlib=1.1.0=pypi_0
graphite2=1.3.14=hc377ac9_1
grpcio=1.59.0=pypi_0
gst-plugins-base=1.14.1=h313beb8_1
gstreamer=1.14.1=h80987f9_1
harfbuzz=4.3.0=he9eebac_1
hdf5=1.12.1=h05c076b_3
icu=68.2=hbdafb3b_0
idna=3.4=py311hca03da5_0
imageio=2.31.4=pypi_0
imgaug=0.4.0=pypi_0
ipykernel=6.25.2=pypi_0
ipython=8.16.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.0=pypi_0
jinja2=3.1.2=py311hca03da5_0
jmespath=1.0.1=pypi_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h80987f9_1
json5=0.9.14=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.1=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.3.1=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.2=pypi_0
jupyter-events=0.7.0=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=2.7.3=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=4.0.6=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.25.0=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.4=py311h313beb8_0
krb5=1.20.1=h69eda48_0
lame=3.100=h1a28f6b_0
lazy-loader=0.3=pypi_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.1.0=hb547adb_0
libbrotlidec=1.1.0=hb547adb_0
libbrotlienc=1.1.0=hb547adb_0
libclang=12.0.0=default_hc321e17_4
libcurl=8.2.1=h3e2b118_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.4=h80987f9_0
libjpeg-turbo=2.0.0=h1a28f6b_0
libllvm12=12.0.1=h93073aa_2
libnghttp2=1.52.0=h62f6fdd_1
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.39=h80987f9_0
libpq=12.15=h02f6b3c_1
libssh2=1.10.0=h02f6b3c_2
libtasn1=4.19.0=h80987f9_0
libtiff=4.5.1=h313beb8_0
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.4=h372ba2a_0
libxslt=1.1.37=h1bd8bc4_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markdown=3.5=pypi_0
markupsafe=2.1.1=py311h80987f9_0
matplotlib=3.7.2=py311hca03da5_0
matplotlib-base=3.7.2=py311h7aedaa7_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.2=pypi_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.3.0=py311hca03da5_0
munkres=1.1.4=pyh9f0ad1d_0
mypy-extensions=1.0.0=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.8.0=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.8=pypi_0
nettle=3.7.3=h84b5d62_1
networkx=3.1=py311hca03da5_0
notebook=7.0.4=pypi_0
notebook-shim=0.2.3=pypi_0
nspr=4.35=hb7217d7_0
nss=3.89.1=h313beb8_0
numexpr=2.8.4=py311h6dc990b_1
numpy=1.25.2=py311he598dae_0
numpy-base=1.25.2=py311hfbfe69c_0
oauthlib=3.2.2=pypi_0
opencv=4.6.0=py311hbae66a1_5
opencv-python=4.7.0.72=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openh264=1.8.0=h98b2900_0
openssl=3.1.3=h53f4e23_0
overrides=7.4.0=pypi_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=py311h7aedaa7_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
patchify=0.2.3=pypi_0
pathspec=0.11.2=pypi_0
patsy=0.5.3=pyhd8ed1ab_0
pcre=8.45=hbdafb3b_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py311h313beb8_1
pip=23.2.1=py311hca03da5_0
pixman=0.40.0=h27ca646_0
platformdirs=3.10.0=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.23.4=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.16.1=pypi_0
pyopenssl=23.2.0=py311hca03da5_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.5=hb885b13_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-json-logger=2.0.7=pypi_0
python-tzdata=2023.3=pyhd8ed1ab_0
pytorch=2.2.0.dev20230921=py3.11_0
pytorch-model-summary=0.1.1=py_0
pytz=2023.3.post1=pyhd8ed1ab_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=py311h80987f9_1
pyzmq=25.1.1=pypi_0
qt-main=5.15.2=ha2d02b5_7
qt-webengine=5.15.9=h2903aaf_7
qtconsole=5.4.4=pypi_0
qtpy=2.4.0=pypi_0
qtwebkit=5.212=h19f419d_5
qudida=0.0.4=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.30.2=pypi_0
requests=2.31.0=py311hca03da5_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.10.3=pypi_0
rsa=4.7.2=pypi_0
s3transfer=0.7.0=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.2.2=py311h313beb8_1
scipy=1.11.1=py311hc76d9b0_0
seaborn=0.12.2=hd8ed1ab_0
seaborn-base=0.12.2=pyhd8ed1ab_0
send2trash=1.8.2=pypi_0
setuptools=68.0.0=py311hca03da5_0
shapely=2.0.1=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h80987f9_0
stack-data=0.6.2=pypi_0
statsmodels=0.14.0=py311hb9f6ed7_0
sympy=1.11.1=py311hca03da5_0
tensorboard=2.15.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.2.0=pyha21a80b_0
tifffile=2023.9.18=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=hb8d0fd4_0
torchaudio=2.2.0.dev20230921=py311_cpu
torchsummary=1.5.1=pypi_0
torchvision=0.17.0.dev20230922=py311_cpu
tornado=6.3.2=py311h80987f9_0
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.10.1=pypi_0
typing_extensions=4.7.1=py311hca03da5_0
tzdata=2023c=h04d1e81_0
uri-template=1.3.0=pypi_0
urllib3=1.26.16=py311hca03da5_0
wcwidth=0.2.8=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.3=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.38.4=py311hca03da5_0
widgetsnbextension=4.0.9=pypi_0
x264=1!152.20180806=h1a28f6b_0
xz=5.4.2=h80987f9_0
yaml=0.2.5=h1a28f6b_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
